Move compute_haaqi to haaqi.py
