+++ clarity/evaluator/haaqi/__init__.py
@@ -1,3 +1,3 @@
-from clarity.evaluator.haaqi.haaqi import haaqi_v1
+from clarity.evaluator.haaqi.haaqi import compute_haaqi, haaqi_v1
 
-__all__ = ["haaqi_v1"]
+__all__ = ["haaqi_v1", "compute_haaqi"]
