+++ recipes/cad1/task1/baseline/evaluate.py
@@ -12,7 +12,7 @@ import pandas as pd
 from omegaconf import DictConfig
 from scipy.io import wavfile
 
-from clarity.evaluator.haaqi import haaqi_v1
+from clarity.evaluator.haaqi import compute_haaqi
 
 logger = logging.getLogger(__name__)
 
@@ -90,34 +90,6 @@ class ResultsFile:
                     str(instruments_scores["right_vocals"]),
                 ]
             )
-
-
-def compute_haaqi(
-    enh_signal: np.ndarray,
-    ref_signal: np.ndarray,
-    audiogram: np.ndarray,
-    audiogram_frequencies: np.ndarray,
-    fs_signal: int,
-) -> float:
-    """Compute HAAQI metric"""
-
-    haaqi_audiogram_freq = [250, 500, 1000, 2000, 4000, 6000]
-    audiogram_adjusted = np.array(
-        [
-            audiogram[i]
-            for i in range(len(audiogram_frequencies))
-            if audiogram_frequencies[i] in haaqi_audiogram_freq
-        ]
-    )
-    score, _, _, _ = haaqi_v1(
-        reference=ref_signal,
-        reference_freq=fs_signal,
-        processed=enh_signal,
-        processed_freq=fs_signal,
-        hearing_loss=audiogram_adjusted,
-        equalisation=1,
-    )
-    return score
 
 
 def set_song_seed(song: str) -> None:
